+++ src/php/CoreBundle/Domain/ClassToCatwalkPackageMigrationHandler.php
@@ -12,20 +12,11 @@ class ClassToCatwalkPackageMigrationHandler
     ) {
         $newClass = "$newNameSpace\\" . $newClassName ?? $className;
         $originalClass = "$originalNameSpace\\$className";
-        $legacyClass = "$originalNameSpace\\Legacy$className";
 
-        if (class_exists($newClass, true)) {
-            if (!self::isProductionEnvironment()) {
-                trigger_error("$originalClass is deprecated, please switch to $newClass");
-            }
-            class_alias($newClass, $originalClass);
-        } else {
-            if (!self::isProductionEnvironment()) {
-                // phpcs:ignore
-                trigger_error("$newClass can not be found. You might be running a version of common library that requires an update of catwalk. Please update to the latest catwalk version!");
-            }
-            class_alias($legacyClass, $originalClass);
+        if (!self::isProductionEnvironment()) {
+            trigger_error("$originalClass is deprecated, please switch to $newClass");
         }
+        class_alias($newClass, $originalClass);
     }
 
     private static function isProductionEnvironment(): bool